+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added default `return503OnClosing: false` option to Fastify
+- Added `SIGTERM` signal handler
+
+### Changed
+
+- Update fastify 2.7.1 -> 2.11.0
+
 ## v2.3.0 - 2019-11-21
 
 ### Changed